test(runtime): split capability recovery regressions
